Move reported_error
