+++ src/tribler-core/tribler_core/components/reporter/reported_error.py
