+++ src/tribler-gui/tribler_gui/tests/test_error_handler.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock, PropertyMock, patch
 
 import pytest
 
-from tribler_common.reported_error import ReportedError
+from tribler_core.components.reporter.reported_error import ReportedError
 from tribler_common.sentry_reporter.sentry_reporter import SentryReporter, SentryStrategy
 
 from tribler_gui.error_handler import ErrorHandler
